feat: errors exceptions

+++ src/modules/users/users.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { UserRepository } from 'src/repositories/users.repository';
@@ -15,14 +15,26 @@ export class UsersService {
   }
 
   findOne(id: string) {
-    return this.usersRepository.findOne(id);
+    const findUser = this.usersRepository.findOne(id);
+    if (!findUser) {
+      throw new NotFoundException('user not found');
+    }
+    return findUser;
   }
 
   update(id: string, updateUserDto: UpdateUserDto) {
+    const findUser = this.usersRepository.findOne(id);
+    if (!findUser) {
+      throw new NotFoundException('user not found');
+    }
     return this.usersRepository.update(id, updateUserDto);
   }
 
   remove(id: string) {
+    const findUser = this.usersRepository.findOne(id);
+    if (!findUser) {
+      throw new NotFoundException('user not found');
+    }
     return this.usersRepository.delete(id);
   }
 }
